Replace complex128 by complex64

+++ tests/test_comparison.py
@@ -200,7 +200,7 @@ class TestComparison:
     def test_n_qubit_circuit(self, wires, lightning_qubit_dev, default_qubit_dev):
         """Test an n-qubit circuit"""
 
-        vec = np.array([1] * (2 ** wires)) / np.sqrt(2 ** wires)
+        vec = np.array([1] * (2 ** wires)) / np.sqrt(2 ** wires, dtype=np.complex64)
         w = qml.init.strong_ent_layers_uniform(2, wires)
 
         def circuit():
@@ -218,4 +218,4 @@ class TestComparison:
 
         default()
         default_state = default_qubit_dev.state
-        assert np.allclose(lightning_state, default_state)
+        assert np.allclose(lightning_state, default_state) # will PASS for atol= 1e-4
